Add a small extension
